requirements.txt
adblockparser==0.7
altgraph==0.17.3
anytree==2.8.0
asn1crypto==1.5.1
astc-decomp-faster==1.1.0
autocommand==2.2.2
base45==0.4.4
bcrypt==4.0.1
beautifulsoup4==4.11.1
biplist==1.0.3
blackboxprotobuf==1.0.1
blinker==1.5
bs4==0.0.1
certifi==2022.9.24
certvalidator==0.11.1
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.1
cheroot==9.0.0
CherryPy==18.8.0
cherrypy-cors==1.6
click==8.1.3
colorama==0.4.6
construct==2.9.45
distro==1.8.0
distro-info==1.0
engineering-notation==0.8.0
et-xmlfile==1.1.0
ExifRead==2.3.2
fasteners==0.18
httpagentparser==1.9.5
httplib2==0.21.0
idna==3.4
importlib-metadata==5.0.0
inflect==6.0.2
ipaddr==2.2.0
ipwhois==1.1.0
jaraco.classes==3.2.3
jaraco.collections==3.8.0
jaraco.context==4.2.0
jaraco.functools==3.5.2
jaraco.text==3.11.0
jeepney==0.8.0
kaitaistruct==0.10
keyring==23.11.0
launchpadlib==1.10.18
lazr.restfulclient==0.14.5
libvmdk-python==20181227
lxml==4.9.1
lz4==4.0.2
Mako==1.2.4
more-itertools==9.0.0
Naked==0.1.32
natsort==8.4.0
netaddr==0.8.0
networkx==2.6.3
nska-deserialize==1.3.2
numpy==1.24.3
oauthlib==3.2.2
opencv-contrib-python==4.7.0.72
openpyxl==3.0.10
oscrypto==1.3.0
packaging==20.1
pandas==1.5.1
paramiko==2.12.0
pathlib2==2.3.5
Pebble==5.0.2
pefile==2022.5.30
PGPy==0.5.4
phonenumbers==8.13.1
Pillow==9.3.0
plistutils==1.0.0
portend==3.1.0
protobuf==3.10.0
psutil==5.9.4
publicsuffixlist==0.7.14
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyBindGen==0.21.0
pycparser==2.21
pycrypto==2.6.1
pycryptodome==3.15.0
pydantic==1.10.2
pydecipher==1.0.0
pygame==2.5.0
pygexf==0.2.2
Pygments==2.13.0
pyinstaller==5.6.2
pyinstaller-hooks-contrib==2022.13
PyJWT==2.6.0
pyliblzfse==0.4.1
pyparsing==3.0.9
PyPDF2==1.28.6
PySimpleGUI==4.16.0
PySocks==1.7.1
python-dateutil==2.8.2
python-debian==0.1.48
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.24
python-pptx==0.6.21
python-whois==0.7.3
pytsk3==20170802
pytz==2022.6
pyxdg==0.28
PyYAML==5.4.1
pyzbar==0.1.9
reportlab==3.6.12
requests==2.28.1
SecretStorage==3.3.3
secure==0.3.0
segno==1.5.2
signify==0.4.0
simplekml==1.3.6
soupsieve==2.3.2.post1
spark-parser==1.8.9
tempora==5.1.0
textdistance==4.5.0
tk==0.1.0
tk-tools==0.16.0
ttkthemes==3.2.2
typing_extensions==4.4.0
uncompyle6==3.8.0
urllib3==1.26.12
xdg==5.1.1
xdis==6.0.4
XlsxWriter==3.0.3
zc.lockfile==2.0
zipp==3.10.0
